+++ src/dao/postgre/ls_loan_closing.rs
@@ -83,7 +83,7 @@ impl Table<LS_Loan_Closing> {
         let value: (Option<BigDecimal>,) = sqlx::query_as(
             r#"
                 SELECT
-                SUM(c."LS_pnl" / POWER(10, cr_asset.decimal_digits)) AS "Total Adjusted Stable Amount"
+                SUM(c."LS_pnl" / POWER(10, cr_asset.decimal_digits)::NUMERIC) AS "Total Adjusted Stable Amount"
                 FROM
                 "LS_Loan_Closing" c
                 LEFT JOIN
@@ -172,7 +172,7 @@ impl Table<LS_Loan_Closing> {
                 repayments AS (
                 SELECT
                     r."LS_contract_id",
-                    SUM(r."LS_payment_amnt_stable" / POWER(10, cr_pay.decimal_digits)) AS total_repaid_usdc
+                    SUM(r."LS_payment_amnt_stable" / POWER(10, cr_pay.decimal_digits)::NUMERIC) AS total_repaid_usdc
                 FROM "LS_Repayment" r
                 INNER JOIN currency_registry cr_pay ON cr_pay.ticker = r."LS_payment_symbol"
                 GROUP BY r."LS_contract_id"
@@ -181,7 +181,7 @@ impl Table<LS_Loan_Closing> {
                 collects AS (
                 SELECT
                     lc."LS_contract_id",
-                    SUM(lc."LS_amount_stable" / POWER(10, cr_col.decimal_digits))::numeric(38,8) AS total_collected_usdc
+                    SUM(lc."LS_amount_stable" / POWER(10, cr_col.decimal_digits)::NUMERIC)::numeric(38,8) AS total_collected_usdc
                 FROM "LS_Loan_Collect" lc
                 INNER JOIN currency_registry cr_col ON cr_col.ticker = lc."LS_symbol"
                 GROUP BY lc."LS_contract_id"
@@ -200,14 +200,14 @@ impl Table<LS_Loan_Closing> {
                 ct.close_ts                                                               AS "LS_timestamp",
                 to_char(ct.close_ts, 'YYYY-MM-DD HH24:MI UTC')                            AS "Close Date UTC",
                 (
-                    (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits))::numeric(38,8)
+                    (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits)::NUMERIC)::numeric(38,8)
                     + COALESCE(r.total_repaid_usdc, 0::numeric(38,8))
                 )::double precision                                                          AS "Sent (USDC, Opening)",
                 COALESCE(c.total_collected_usdc, 0::numeric(38,8))::double precision          AS "Received (USDC, Closing)",
                 (
                     COALESCE(c.total_collected_usdc, 0::numeric(38,8))
                     - (
-                        (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits))::numeric(38,8)
+                        (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits)::NUMERIC)::numeric(38,8)
                         + COALESCE(r.total_repaid_usdc, 0::numeric(38,8))
                     )
                 )::double precision                                                           AS "Realized PnL (USDC)"
@@ -258,7 +258,7 @@ impl Table<LS_Loan_Closing> {
                 repayments AS (
                 SELECT
                     r."LS_contract_id",
-                    SUM(r."LS_payment_amnt_stable" / POWER(10, cr_pay.decimal_digits)) AS total_repaid_usdc
+                    SUM(r."LS_payment_amnt_stable" / POWER(10, cr_pay.decimal_digits)::NUMERIC) AS total_repaid_usdc
                 FROM "LS_Repayment" r
                 INNER JOIN currency_registry cr_pay ON cr_pay.ticker = r."LS_payment_symbol"
                 GROUP BY r."LS_contract_id"
@@ -268,7 +268,7 @@ impl Table<LS_Loan_Closing> {
                 collects AS (
                 SELECT
                     lc."LS_contract_id",
-                    SUM(lc."LS_amount_stable" / POWER(10, cr_col.decimal_digits))::numeric(38,8) AS total_collected_usdc
+                    SUM(lc."LS_amount_stable" / POWER(10, cr_col.decimal_digits)::NUMERIC)::numeric(38,8) AS total_collected_usdc
                 FROM "LS_Loan_Collect" lc
                 INNER JOIN currency_registry cr_col ON cr_col.ticker = lc."LS_symbol"
                 GROUP BY lc."LS_contract_id"
@@ -286,7 +286,7 @@ impl Table<LS_Loan_Closing> {
                     o."LS_contract_id"                            AS position_id,
                     -- Sent at opening: normalized collateral + total repayments (USDC)
                     (
-                    (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits))::numeric(38,8)
+                    (o."LS_cltr_amnt_stable" / POWER(10, cr_cltr.decimal_digits)::NUMERIC)::numeric(38,8)
                     + COALESCE(r.total_repaid_usdc, 0::numeric(38,8))
                     )                                               AS sent_open_usdc,
                     -- Received at closing: normalized collects (USDC). Zero if liquidated without collects.
